refac(test): deduplicate pytest_collection_modifyitems
